pull getDependencies into separate file
